pyproject.toml
[tool.poetry]
name = "aggregator-service"
version = "0.1.0"
description = ""
authors = ["thee <theekshana.sandaru@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.35.89"
psycopg2-binary = "^2.9.10"
redshift-connector = "^2.1.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
